added chat
